requirements.txt
fastapi
pyrogram
uvicorn
python-dotenv

